requirements.txt
fastapi
uvicorn
python-multipart
openai
pymupdf
pytesseract
pillow
python-dotenv
